requirements.txt
streamlit
requests
pandas
python-dotenv
scikit-learn
joblib
